\encoding{utf8}
\name{maritimes.coords}
\alias{maritimes.coords}
\alias{geofd.maritimes.coords}
\docType{data}
\title{Coordinates of the sites referred by maritimes.data}
\description{
  The geographical coordinates in decimal degrees of 35 weather stations.
}
\usage{data(maritimes.coords)}
\format{
 A matrix with the coordinates of 35 weather stations.
}
\source{
  The coordinates were obtained from the database of geographic coordinate information \url{http://www.tageo.com}
}
\keyword{datasets}
