\name{maritimes.data}
\alias{maritimes.data}
\alias{geofd.maritimes.data}
\docType{data}
\title{Maritime provinces temperatures}
\description{
  Temperature measurements recorded at 35 weather stations located in the Maritime Provinces over a region of Canada consisting of three provinces: Nova Scotia (NS), New Brunswick (NB), and Prince Edward Island (PEI).
}
\usage{data(maritimes)}
\format{
  A matrix with 365 observations on 35 sites.
}
\details{
  This data set contains information of daily temperatures averaged over the years 1960 to 1994 (February 29th combined with February 28th)
}
\source{
The data for each station were obtained from the Meteorological Service of Canada \url{http://www.climate.weatheroffice.ec.gc.ca/climateData/}
}
\keyword{datasets}
