\encoding{utf8}
\name{geofd-package}
\alias{geofd-package}
\alias{geofd}
\docType{package}
\title{
Spatial prediction for function value data
}
\description{
Prediction for function value spatial data using kriging based methods
}
\details{
\tabular{ll}{
Package: \tab geofd\cr
Type: \tab Package\cr
Version: \tab 0.4.6\cr
LazyLoad: \tab yes\cr
LazyData: \tab yes\cr
License: \tab GPL-2\cr
Date: \tab 2010-06-16\cr
Depends: \tab R (>= 2.10.0), fda, geoR, tkrplot\cr
Built: \tab R 2.11.1; ; 2010-08-09 17:22:47 UTC; unix\cr
}

Index:
\preformatted{
geofd-package           Spatial prediction for function value data
l2.norm                 Calculates L2 norm among functions
maritimes.avg           Moncton averages
maritimes.coords        Coordinates of the sites referred by
                        maritimes.data
maritimes.data          Maritime provinces temperatures
okfd                    Function for doing Ordinary Kriging for
                        function-value Data
okfd.cv                 Function for doing Cross-Validation analysis
                        for Ordinary Kriging for function-value data
plot.geofd              Plot Trace Variogram and adjusted models
fit.tracevariog         Variogram Estimation
trace.variog            Empirical Variograms for function-value data
}
}
\author{
Ramon Giraldo <rgiraldoh@unal.edu.co>, Pedro Delicado <pedro.delicado@upc.edu>, Jorge Mateu <mateu@mat.uji.es>

Maintainer: Andres Perez <afperesm@unal.edu.co>
}
\references{
  Giraldo, R. (2009) \emph{Geostatistical Analysis of Functional Data.} Ph.D. thesis. Universitat Politècnica de Catalunya.
}
\keyword{spatial,datagen}
